requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dash==0.36.0
dash-bootstrap-components==0.3.4
dash-core-components==0.43.0
dash-daq==0.1.0
dash-html-components==0.13.5
dash-renderer==0.18.0
dash-table==3.1.11
dash.ly==0.17.3
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
nbformat==4.4.0
numpy==1.16.1
pandas==0.24.1
plotly==3.6.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
requests-cache==0.4.13
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
